refactor(arch): break the last circular dependency (#601)
The agent/executor/handler/manager + ssh/pool + nodes + executor 6-node
cycle was the last remaining one. Other cycles listed in #601 (mcp/
tokens, stackInstall/credentialsManifest, config/registry/migrations)
have already been broken by other refactors.

The cycle existed because `nodes.ts:verifyNodeConnection` did
\`await import('./executor')\` to run \`podman info\` remotely.
\`nodes.ts\` is a pure config-IO module (read/write nodes.json,
manage default-node pointer); the executor runtime had no business
hanging off it.

Fix: extracted \`verifyNodeConnection\` to a new
\`src/lib/nodes/verify.ts\` that owns the runtime-executor coupling.
\`nodes.ts\` now has zero imports into the agent/executor graph.

Also:
- Exported \`normalizeName\` from \`nodes.ts\` so \`verify.ts\` does
  the same case-insensitive node lookup.
- Removed the \`pathNot\` exemption from \`no-circular\` in
  \`.dependency-cruiser.cjs\`. Any new cycle now fails CI.
- Updated \`docs/ARCHITECTURE_INVARIANTS.md\` row: 6 known → 0.

Closes #601.
